third_party/blink/web_tests/webexposed/feature-policy-features.html

<!DOCTYPE html>
<pre id="output"></pre>
<script src="../resources/feature-policy-features-listing.js"></script>
<script>

// This test verifies that in a normal (non-isolated) context, a set of
// expected permissions policies appears. There is an isolated version of this
// test at:
// //third_party/blink/web_tests/wpt_internal/isolated-permissions-policy/

if (window.testRunner) {
  testRunner.dumpAsText();
}

function print(text) {
  output.textContent += text + '\n';
}

print('This test documents Blink\'s web-exposed policy-controlled features.');
print('All changes to this list should go through Blink\'s feature review process: http://www.chromium.org/blink#new-features');
print('');
featurePolicyFeaturesListing(false /* platformSpecific */, print);
</script>
