third_party/blink/web_tests/virtual/stable/webexposed/feature-policy-features-expected.txt

This test documents Blink's web-exposed policy-controlled features.
All changes to this list should go through Blink's feature review process: http://www.chromium.org/blink#new-features

accelerometer
autoplay
camera
document-domain
encrypted-media
fullscreen
geolocation
gyroscope
magnetometer
microphone
midi
payment
picture-in-picture
sync-xhr
usb
xr-spatial-tracking
